Simplify cover highlighting and remove it from cache.
Resize and filtering operations seem to be async, and it caused problems
in placeholder thumbnails which only rendered 50% when dragging. Probs
ceased once we bypassed cache and highlight.

Now the cache is simpler and therefore more performant, and the new opacity
based "highlight" is quite alright and looks visually better than
the previous idea.